+++ RELEASE-NOTES.md
@@ -5,6 +5,8 @@ Version 0.2.0
 -------------
 
 New features:
+* Support for serializing Wikibase data in RDF (as illustrated in new example);
+  see http://korrekt.org/page/Introducing_Wikidata_to_the_Linked_Data_Web for details
 * Simplified code for dump file processing: new helper class DumpProcessingController
 * Support for resolving site links, based on information from the sites table dump
   (as demonstrated in a new example program)
@@ -12,6 +14,7 @@ New features:
 * Support for serializing Wikibase data in JSON (as illustrated in new example)
 
 Bug fixes:
+* Support changed Wikimedia dump HTML page format, which caused download to fail (Issue #70)
 * Support processing of property documents when parsing dumps (Issue #67)
 * Support SomeValueSnak and NoValueSnak in references (Issue #44)
 * Use correct site links when importing data from dumps (Issue #37)
@@ -21,7 +24,9 @@ Incompatible changes:
 * The processing of dumpfiles was simplified, using a new class DumpProcessingController.
   The former method WmfDumpFileManager#processRecentRevisionDumps() was replaced by
   DumpProcessingController#processAllRecentRevisionDumps(). See the examples for example
-  code. 
+  code.
+* Dump files no longer support the retrieval of the maximal revision id, since this
+  information is no longer published for the main dumps on the Wikimedia site.
 
 
 Version 0.1.0